Extract interface for ItemCountPage
